X-Clone-Backend-FastAPI/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
SQLAlchemy
alembic
PyJWT
passlib[bcrypt]
python-multipart
aiofiles
pymysql


